Fix: Move ttsLanguage variable outside usePlayer in usePlayer.ts, otherwise it is undefined on MainPage.
This is due to a change where we now call setTtsLanguage from LanguageSelectionPage.vue. Since usePlayer.ts contains a composable, importing it in both LanguageSelectionPage.vue and MainPage.vue breaks the setting of the TTS language, if ttsLanguage is not outside usePlayer. Using a composable here turns out to be a major source of bugs, since we currently don't have test coverage for affected files.