fix: Quick fix to make TTS work
For some TTS files, only some of the lines are spoken. I am working on a fix, but it will take a while. Until then, we just set the exact time of each line to 0, meaning it will be spoken at soon as the client receives it. This is not ideal, but better than nothing.